// src/app/models/product.model.ts
export class Product {
  id: number;
  vendorId: number;
  categoryId: number;
  categoryName: string;
  subCategoryId: number;
  subCategoryName: string;
  name: string;
  brand: string;
  location: string;
  status: string;
  color:string;
  size:string;
  description: string;
  fullDescription: string;
  features: string;
  deliveryDays: number;
  paymentOption: string;
  price: number;
  ratings: number;
  reviews: number;
  stock: string;
  images: any;
  source: string;
  sizes: string;
  thumbnail: any;
  remark: string;
}